Split up files.py
