+++ setup.cfg
@@ -109,7 +109,6 @@ ignore = E203,W503
 exclude =
     *sphinx*
     dandi/externals/*
-    */__init__.py
     .tox/
     _version.py
     versioneer.py
